Extract and reuse the fixOffset method - Create and reuse a method to parse incoming raw iso8601 datetimes and then converting them into Date objects
