+++ src/main/java/org/opendatakit/common/persistence/engine/EngineUtils.java
@@ -19,6 +19,7 @@ package org.opendatakit.common.persistence.engine;
 
 import static java.time.ZoneId.systemDefault;
 import static java.time.format.DateTimeFormatter.ISO_OFFSET_DATE_TIME;
+import static org.opendatakit.aggregate.submission.type.jr.JRTemporalUtils.parseDate;
 
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
@@ -192,7 +193,7 @@ public class EngineUtils {
         if (v == null) {
           value = null;
         } else {
-          value = Date.from(OffsetDateTime.parse(fixOffset(v)).toInstant());
+          value = parseDate(v);
         }
         break;
       }
@@ -200,11 +201,6 @@ public class EngineUtils {
         throw new IllegalStateException("datatype not handled");
     }
     return value;
-  }
-
-  private static String fixOffset(String raw) {
-    char thirdCharFromTheEnd = raw.charAt(raw.length() - 3);
-    return thirdCharFromTheEnd == '+' || thirdCharFromTheEnd == '-' ? raw + ":00" : raw;
   }
 
 }
